# swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Bot Ultra
  description: >
    Nodejs BackEnd for Bot Ultra.
...
schemes:
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /api/admin/user:
    post:
      summary: Created New User
      tags:
        - user
      security:
        - Bearer: []

      parameters:
        - in: body
          name: body
          description: Register need object
          required: false
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: "User Result"
        "403":
          description: "You do not have necessary permissions for the resource"

    get:
      summary: get all users
      tags:
        - user
      security:
        - Bearer: []
      parameters:
        - name: pageindex
          in: query
          description: current pageindex
          required: false
          type: number
        - name: pagesize
          in: query
          description: total page size
          required: false
          type: number
      responses:
        "200":
          description: "User Result"
        "403":
          description: "You do not have necessary permissions for the resource"

    put:
      summary: update user
      tags:
        - user
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          description: Register need object
          required: false
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: "User Result"
        "403":
          description: "You do not have necessary permissions for the resource"
  /api/admin/user/{id}:
    get:
      summary: Get user
      tags:
        - user
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: string
          required: true
      responses:
        "200":
          description: "User Result"
        "403":
          description: "You do not have necessary permissions for the resource"
    delete:
      summary: delete user
      tags:
        - user
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: string
          required: true
      responses:
        "200":
          description: "User Result"
        "403":
          description: "You do not have necessary permissions for the resource"
  /api/zyngabot:
    post:
      summary: Created New Zynga Bot
      tags:
        - zynga
      security:
        - Bearer: []

      parameters:
        - in: body
          name: body
          description: Register need object
          required: false
          schema:
            $ref: "#/definitions/ZyngaBot"
      responses:
        "200":
          description: "zynga Result"
        "403":
          description: "You do not have necessary permissions for the resource"

    get:
      summary: get all zynga for user
      tags:
        - zynga
      security:
        - Bearer: []
      responses:
        "200":
          description: "zynga Result"
        "403":
          description: "You do not have necessary permissions for the resource"

    put:
      summary: update zynga
      tags:
        - zynga
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          description: Register need object
          required: false
          schema:
            $ref: "#/definitions/ZyngaBot"
      responses:
        "200":
          description: "zynga Result"
        "403":
          description: "You do not have necessary permissions for the resource"
  /api/zyngabot/{id}:
    get:
      summary: get  zyngabot for id
      tags:
        - zynga
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: string
          required: true
      responses:
        "200":
          description: "zynga Result"
        "403":
          description: "You do not have necessary permissions for the resource"
    delete:
      summary: delete zynga
      tags:
        - zynga
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: string
          required: true
      responses:
        "200":
          description: "zynga Delete Result"
        "403":
          description: "You do not have necessary permissions for the resource"
  /api/zyngabot/images/{id}:
    get:
      summary: get  zyngabot image for id
      tags:
        - zynga
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: string
          required: true
      responses:
        "200":
          description: "zynga Result"
        "403":
          description: "You do not have necessary permissions for the resource"
    delete:
      summary: delete zynga
      tags:
        - zynga
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          type: string
          required: true
      responses:
        "200":
          description: "zynga Delete Result"
        "403":
          description: "You do not have necessary permissions for the resource"
  /api/zyngabot/bot/command:
    post:
      summary: Post Bot Command
      tags:
        - zynga
      security:
        - Bearer: []

      parameters:
        - in: body
          name: body
          description: Post Bot Command
          required: false
          schema:
            $ref: "#/definitions/ZyngaCommand"
      responses:
        "200":
          description: "Task Result"
        "403":
          description: "You do not have necessary permissions for the resource"
...
definitions:
...
  User:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      roles:
        type: array
        items:
          type: string
      lastloginip:
        type: string
  ZyngaBot:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      createddate:
        type: string
        format: date-time
      status:
        type: number
      isactive:
        type: boolean
      loginusername:
        type: string
      loginpassword:
        type: string
    example:
      name: deneme zynga botu
      description: demoilhan
      status: 0
      isactive: true
  ZyngaCommand:
    type: object
    properties:
      destbotid:
        type: string
      sourceusername:
        type: string
      sourcepassword:
        type: string
      sendblance:
        type: string
    example:
      destbotid: ""
      sourceusername: ""
      sourcepassword: ""
      sendblance: ""
...